pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Web application for analyzing URLs SEO protection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flake8>=7.2.0",
    "flask>=3.1.0",
    "gunicorn>=23.0.0", 
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
]
